requirements.txt
environs==7.1.0
flask==1.1.1
flask-graphql==2.0.1
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
psycopg2==2.7.3
python-dotenv==0.7.1
sqlalchemy==1.3.12
